+++ android/src/main/java/vn/hunghd/flutterdownloader/DownloadedFileProvider.java
@@ -1,6 +1,6 @@
 package vn.hunghd.flutterdownloader;
 
-import android.support.v4.content.FileProvider;
+import androidx.core.content.FileProvider;
 
 public class DownloadedFileProvider extends FileProvider {
 }